feat: add channel tabs
