+++ app/src/main/java/org/schabi/newpipe/App.java
@@ -17,6 +17,7 @@ import org.acra.config.CoreConfigurationBuilder;
 import org.schabi.newpipe.error.ReCaptchaActivity;
 import org.schabi.newpipe.extractor.NewPipe;
 import org.schabi.newpipe.extractor.downloader.Downloader;
+import org.schabi.newpipe.extractor.services.youtube.YoutubeParsingHelper;
 import org.schabi.newpipe.ktx.ExceptionUtils;
 import org.schabi.newpipe.settings.NewPipeSettings;
 import org.schabi.newpipe.util.Localization;
@@ -89,6 +90,10 @@ public class App extends Application {
         NewPipe.init(getDownloader(),
             Localization.getPreferredLocalization(this),
             Localization.getPreferredContentCountry(this));
+
+        // TODO: REMOVE THIS, only for testing
+        YoutubeParsingHelper.setVisitorData("CgtOa256ckVkcG5YVSi7-c6aBg%3D%3D");
+
         Localization.initPrettyTime(Localization.resolvePrettyTime(getApplicationContext()));
 
         StateSaver.init(this);
