+++ app/src/main/java/org/schabi/newpipe/settings/tabs/Tab.java
@@ -19,7 +19,7 @@ import org.schabi.newpipe.extractor.NewPipe;
 import org.schabi.newpipe.extractor.StreamingService;
 import org.schabi.newpipe.extractor.exceptions.ExtractionException;
 import org.schabi.newpipe.fragments.BlankFragment;
-import org.schabi.newpipe.fragments.list.channel.ChannelFragment;
+import org.schabi.newpipe.fragments.list.channel.ChannelVideosFragment;
 import org.schabi.newpipe.fragments.list.kiosk.DefaultKioskFragment;
 import org.schabi.newpipe.fragments.list.kiosk.KioskFragment;
 import org.schabi.newpipe.fragments.list.playlist.PlaylistFragment;
@@ -432,8 +432,8 @@ public abstract class Tab {
         }
 
         @Override
-        public ChannelFragment getFragment(final Context context) {
-            return ChannelFragment.getInstance(channelServiceId, channelUrl, channelName);
+        public ChannelVideosFragment getFragment(final Context context) {
+            return ChannelVideosFragment.getInstance(channelServiceId, channelUrl, channelName);
         }
 
         @Override
