+++ app/src/main/res/menu/menu_channel_videos.xml
@@ -1,0 +1,14 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".fragments.list.impl.ChannelFragment">
+
+    <item
+        android:id="@+id/menu_item_notify"
+        android:checkable="true"
+        android:orderInCategory="1"
+        android:title="@string/get_notified"
+        android:visible="false"
+        app:showAsAction="never"
+        tools:visible="true" />
+</menu>
